[BACK] fix: rename babel to backend
