+++ backend/src/main/java/ru/itmo/backend/BabelApplication.java
@@ -1,4 +1,4 @@
-package ru.itmo.babel;
+package ru.itmo.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
