+++ backend/src/main/java/ru/itmo/backend/dto/AnalyzeRequest.java
@@ -1,3 +1,3 @@
-package ru.itmo.babel.dto;
+package ru.itmo.backend.dto;
 
 public record AnalyzeRequest(String url){}
