+++ backend/src/main/java/ru/itmo/backend/entity/GitRepositoryEntity.java
@@ -1,4 +1,4 @@
-package ru.itmo.babel.entity;
+package ru.itmo.backend.entity;
 
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
