+++ backend/src/test/java/ru/itmo/backend/BabelApplicationTests.java
@@ -1,6 +1,5 @@
-package ru.itmo.babel;
+package ru.itmo.backend;
 
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
